chore: move dan signing to standalone helper
